refactor: move search, external-skills, plugin-ops to SDK (Phase 6)
Move the final 3 library modules to SDK: search.ts -> catalog/search.ts,
external-skills.ts -> providers/local/external.ts, plugin-ops.ts ->
context/plugin/ops.ts. CLI lib/ now has 4 non-shim TypeScript files.
